third_party/WebKit/LayoutTests/fast/scrolling/overflow-scrollability-expected.txt

This tests that scrollable areas with the appropriate overflow mode set are in fact scrollable by the user.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
===overflow:scroll with overflow
PASS div.scrollTop is window.pixelsPerWheelTick
PASS div.scrollLeft is window.pixelsPerWheelTick
===overflow:auto with overflow
PASS div.scrollTop is window.pixelsPerWheelTick
PASS div.scrollLeft is window.pixelsPerWheelTick
===overflow:hidden with overflow
PASS div.scrollTop is 0
PASS div.scrollLeft is 0
===overflow:overlay with overflow
PASS div.scrollTop is window.pixelsPerWheelTick
PASS div.scrollLeft is window.pixelsPerWheelTick
===overflow:scroll without overflow
PASS div.scrollTop is 0
PASS div.scrollLeft is 0
===overflow:auto without overflow
PASS div.scrollTop is 0
PASS div.scrollLeft is 0
===overflow:hidden without overflow
PASS div.scrollTop is 0
PASS div.scrollLeft is 0
===overflow:overlay without overflow
PASS div.scrollTop is 0
PASS div.scrollLeft is 0
scroll with overflow
Content
auto with overflow
Content
hidden with overflow
Content
overlay with overflow
Content
scroll without overflow
Content
auto without overflow
Content
hidden without overflow
Content
overlay without overflow
Content
